test: share screenshot navigation helpers
